Problèmes chargement et relecture json fixés
-Valeurs can en float
-Ordre des voies adc changée pour correspondre au rack
-Json protégé par i2c
-adresse i2c1 pour mémoire extérieur
-Structure json en majuscule
-Buffer uart passe à 512 byte

+++ Integrationcppstatemachine/Src/Can.cpp
@@ -99,7 +99,7 @@ uint16_t Can::readADC(){
 
 
 float Can::readADCImoy(int nb){
-	uint16_t value = 0;
+	float value = 0;
 	float moyenne = 0;
 
 	if(this->rack.getRackadress() == NULL){
